[AIRFLOW-3474] Move SlaMiss out of models.py
